config: Update the NicClusterPolicy name
Turns out that the NicClusterPolicy CR can only be called `nic-cluster-policy`.
Otherwise the operator will ignore the CR. This commit updates the name of the
NicClusterPolicy CR in the configs from `aks-nic-cluster-policy` to
`nic-cluster-policy`.

Here is the status message on the custom-resource that the operator throws when
the CR name is not `nic-cluster-policy`:

```yaml
status:
  reason: Unsupported NicClusterPolicy instance aks-nic-cluster-policy. Only
    instance with name nic-cluster-policy is supported
  state: ignore
```

Signed-off-by: Suraj Deshmukh <suraj.deshmukh@microsoft.com>