tests: Convert configs to helm chart
Kustomizate was getting cumbersome to manage and maintain. This commit converts
the test configs to helm chart for better organization and management. Helm also
gives you an ability to template, add conditionals, etc.

Instead of running a full test suite, allow users to run tests in the scenario
they want and the sub-test they want.