nicclusterpolicy: Add env var OFED_BLACKLIST_MODULES_FILE
This commit adds an env var
`OFED_BLACKLIST_MODULES_FILE=/host/etc/modprobe.d/blacklist-ofed-modules.conf`.

This is done to avoid the race condition where the driver container tries to
reload the module but cannot because the host loads the modules shipped with the
kernel.

This workaround ensures that the blacklist file is present on the host, so that
the host does not interfere with the module load process.

Signed-off-by: Suraj Deshmukh <suraj.deshmukh@microsoft.com>