renovate: Add gpu-operator and network-operator configs
This commit adds configuration to update GPU-Operator and Network-Operator. The
configuration is identical to each other. The obvious differences are the fields
`depNameTemplate` and `fileMatch`.

Renovate refers to the `index.yaml` file here to identify the versions of the
operators: https://helm.ngc.nvidia.com/nvidia/index.yaml. GPU-operator and
Network-Operator both publish the versions differently, GPU-operator appends `v`
before the semver and Network-Operator does not.

Here is the snippet of `gpu-operator`:

```yaml
  gpu-operator:
  - apiVersion: v2
    appVersion: v25.3.0
    ...
    urls:
    - charts/gpu-operator-v25.3.0.tgz
    version: v25.3.0
```

And here is the snippet of `network-operator`:

```yaml
  network-operator:
  appVersion: v25.4.0
  - apiVersion: v2
    ...
    urls:
    - charts/network-operator-25.4.0.tgz
    version: 25.4.0
```

Hence the non-obvious difference is that the `matchStrings` for
`network-operator` has `v` prefix added to the semver.

Signed-off-by: Suraj Deshmukh <suraj.deshmukh@microsoft.com>